inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onnxruntime>=1.22.1",
    "pycapnp>=2.2.0",
    "tokenizers>=0.22.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.75.0",
]
